Move:レイアウト関連のコンポーネントファイルを移動 - app-sidebar.tsxを移動 - page-header.tsxを移動 - ヘッダーの下ボーダーを除外
